requirements.txt
# Standard Imports
python-dotenv
requests
